% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceFromFEV1-package.R
\docType{package}
\name{raceFromFEV1-package}
\alias{raceFromFEV1}
\alias{raceFromFEV1-package}
\title{raceFromFEV1: Determines the race variable used to calculate a race-based FEV1. For research use only.}
\description{
Allows to back-calculate the race of the patient from percent predicted FEV1 values when using a race-based spirometry reference equations.
}
\author{
\strong{Maintainer}: Amin Adibi \email{adibi@alumni.ubc.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
